+++ frontend/package.json
@@ -28,7 +28,7 @@
   "scripts": {
     "start": "react-scripts start",
     "test": "react-scripts test",
-    "eject": "react-scripts eject",
+    "eject": "NODE_PATH=src/ react-scripts eject",
     "build": "react-scripts build"
   },
   "eslintConfig": {
